# backend/docs/swagger.yaml
basePath: /
definitions:
...
  oauth2.TokenRequest:
    properties:
      client_id:
        type: string
      client_secret:
        type: string
      code:
        type: string
      code_verifier:
        type: string
      grant_type:
        type: string
      redirect_uri:
        type: string
      refresh_token:
        type: string
    required:
    - client_id
    - grant_type
    type: object
host: localhost:8080
info:
  contact:
    email: support@noraneko-id.com
    name: Noraneko ID Support
    url: https://noraneko-id.com/support
  description: プライベートサービス開発者向けIDaaS（Identity as a Service）API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://noraneko-id.com/terms
  title: Noraneko ID API
  version: "1.0"
paths:
...
  /oauth2/authorize:
    get:
      description: OAuth2認可コードフローの認可エンドポイントです。PKCEをサポートしています。
      parameters:
      - description: レスポンスタイプ
        enum:
        - code
        in: query
        name: response_type
        required: true
        type: string
      - description: クライアントID
        in: query
        name: client_id
        required: true
        type: string
      - description: リダイレクトURI
        in: query
        name: redirect_uri
        required: true
        type: string
      - default: openid
        description: スコープ
        in: query
        name: scope
        type: string
      - description: 状態パラメータ
        in: query
        name: state
        type: string
      - description: PKCEコードチャレンジ
        in: query
        name: code_challenge
        type: string
      - default: S256
        description: PKCEチャレンジメソッド
        enum:
        - S256
        in: query
        name: code_challenge_method
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Redirect to authorization page or callback URL
        "400":
          description: リクエストエラー
          schema:
            additionalProperties: true
            type: object
      summary: OAuth2認可エンドポイント
      tags:
      - OAuth2
  /oauth2/client-info/{client_id}:
    get:
      description: OAuth2クライアントの公開情報を取得します（認可画面用）
      parameters:
      - description: クライアントID
        in: path
        name: client_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: クライアント情報
          schema:
            additionalProperties: true
            type: object
        "404":
          description: クライアントが見つかりません
          schema:
            additionalProperties: true
            type: object
      summary: クライアント情報取得
      tags:
      - OAuth2
  /oauth2/revoke:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: アクセストークンまたはリフレッシュトークンを無効化します。
      parameters:
      - description: 無効化するトークン
        in: formData
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 無効化成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: リクエストエラー
          schema:
            additionalProperties: true
            type: object
      summary: OAuth2トークン無効化エンドポイント
      tags:
      - OAuth2
  /oauth2/token:
    post:
      consumes:
      - application/json
      description: 認可コードやリフレッシュトークンを使ってアクセストークンを取得します。
      parameters:
      - description: トークンリクエスト
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/oauth2.TokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: アクセストークンレスポンス
          schema:
            additionalProperties: true
            type: object
        "400":
          description: リクエストエラー
          schema:
            additionalProperties: true
            type: object
        "401":
          description: クライアント認証エラー
          schema:
            additionalProperties: true
            type: object
      summary: OAuth2トークンエンドポイント
      tags:
      - OAuth2
  /oauth2/userinfo:
    get:
      description: アクセストークンを使ってユーザー情報を取得します。
      produces:
      - application/json
      responses:
        "200":
          description: ユーザー情報
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 認証エラー
          schema:
            additionalProperties: true
            type: object
        "404":
          description: ユーザーが見つからない
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: OAuth2ユーザー情報エンドポイント
      tags:
      - OAuth2
securityDefinitions:
  BearerAuth:
    description: Bearer Token認証。「Bearer 」の後にアクセストークンを指定してください。
    in: header
    name: Authorization
    type: apiKey
  SessionAuth:
    description: セッション認証。session_tokenクッキーを使用します。
    in: header
    name: Cookie
    type: apiKey
swagger: "2.0"